# raspyxauth/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation for authentication/authorization API service",
        "title": "JWT authentication/authorization service",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/auth/api",
    "paths": {
...
        "/v1/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Getting a list of all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Getting users",
                "responses": {
                    "200": {
                        "description": "User data has been successfully received",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ResponseAPI"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creating a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Creating a user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User successfully created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ResponseAPI"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request format",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    },
                    "409": {
                        "description": "Conflict when creating user",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    }
                }
            }
        },
        "/v1/users/accesslevel/{access_level}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Getting information about users whose access level is lower or equal to the specified one",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получение пользователей по уровню доступа",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User access level",
                        "name": "access_level",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User data has been successfully received",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ResponseAPI"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "User was not found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    }
                }
            }
        },
        "/v1/users/username/{username}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Getting information about a user by their username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Getting a user by username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User data has been successfully received",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ResponseAPI"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.User"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "User was not found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    }
                }
            }
        },
        "/v1/users/uuid/{uuid}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Getting information about a user by their UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Getting a user by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User data has been successfully received",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.ResponseAPI"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "User was not found",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseAPI"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AddUserRequest": {
            "type": "object",
            "properties": {
                "access_level": {
                    "type": "integer",
                    "example": 50
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "username": {
                    "type": "string",
                    "example": "username"
                },
                "uuid": {
                    "type": "string",
                    "example": "4b6c34bd-01f0-4fbe-be65-217984b3e33d"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseAPI": {
            "type": "object",
            "properties": {
                "errors": {
                    "$ref": "#/definitions/models.ErrorResponse"
                },
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "result": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "access_level": {
                    "type": "integer",
                    "example": 100
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "username": {
                    "type": "string",
                    "example": "username"
                },
                "uuid": {
                    "type": "string",
                    "example": "4b6c34bd-01f0-4fbe-be65-217984b3e33d"
                }
            }
        },
        "models.UserResponse": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string",
                    "example": "4b6c34bd-01f0-4fbe-be65-217984b3e33d"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}